user authentication

+++ FastAPI/BackendBasics/SQL/routers/todos.py
@@ -5,6 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException, Path
 from models import Todos
 from database import Sessionlocal
 from sqlalchemy.orm import Session
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -17,6 +18,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel): #validation
     title: str = Field(min_length=3)
@@ -25,28 +27,38 @@ class TodoRequest(BaseModel): #validation
     complete: bool
 
 
-@router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency): #Annotated[Session, Depends(get_db)] significa: "esse parâmetro é do tipo Session, e seu valor vem de Depends(get_db)". O FastAPI lê isso, executa get_db() automaticamente antes de chamar read_all, e injeta a sessão no parâmetro db
-    return db.query(Todos).all()
+@router.get("/", status_code=status.HTTP_200_OK) #Mostra todas as tarefas de acordo com o id do cliente autenticado
+async def read_all(user: user_dependency, db: db_dependency): #Annotated[Session, Depends(get_db)] significa: "esse parâmetro é do tipo Session, e seu valor vem de Depends(get_db)". O FastAPI lê isso, executa get_db() automaticamente antes de chamar read_all, e injeta a sessão no parâmetro db
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
+
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo not found.')
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request:TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request:TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update(db: db_dependency,
+async def update(user: user_dependency, db: db_dependency,
                  todo_request: TodoRequest,
                  todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first() #fetch
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first() #fetch
     if todo_model is None: #Error status code
         raise HTTPException(status_code=404, detail='Todo not found')
     #updates
@@ -59,10 +71,12 @@ async def update(db: db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None: #Error status code
+        raise HTTPException(status_code=404, detail='Todo not found')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
